LPS-161984 Extract AggregationFormBase from objectFieldFormBase to its own component
